ozone: caca: Fix the build
TEST=build
BUG=none
NOTRY=true

Review URL: https://codereview.chromium.org/247933003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@265676 0039d316-1c4b-4281-b951-d872f2087c98

+++ ui/ozone/platform/dri/chromeos/native_display_delegate_dri.cc
@@ -14,7 +14,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 namespace ui {
 
 namespace {
-const int kMaxDisplayCount = 2;
+const size_t kMaxDisplayCount = 2;
 }  // namespace
 
 NativeDisplayDelegateDri::NativeDisplayDelegateDri(
